requirements.txt
beautifulsoup4==4.14.3
pymupdf==1.27.2
numpy==2.4.3
protobuf==7.34.1
python-dotenv==1.2.2
python-docx==1.2.0
python-pptx==1.0.2
requests==2.32.5
streamlit==1.55.0
google-genai==1.66.0
